HASTA V 1.1.5

+++ PIGGISWS/Interfaces/IMarcacionService.cs
@@ -6,5 +6,5 @@ public interface IMarcacionService
 {
     Task<ServiceResponse<object>> CreateMarcacionAsync(Tmp_Marcacion_Agente marcacion_Agente);
     Task<ServiceResponse<object>> GetHoraAsync();
-    Task<ServiceResponse<object>> GetMarcacionesxAgenteAsync(int agente);
+    Task<ServiceResponse<object>> GetMarcacionesxAgenteAsync(decimal agente);
 }
